Bug fixes in while loops
This commit is about fixing bugs over implementing features. See commit history document for more specific information.
- Fixed a parsing bug that created problems with empty statement blocks in while loops. All empty statement blocks now generate compiler errors (in while loops, conditionals, and definitions)
- Fixed a bug in while loops contained within local scopes; the compiler did not pass max_offset into a fetch_value call, which resulted in the compiler _thinking_ that addresses were safe that were not guaranteed to be, which could cause stack corruption.
- Added explicit pass statements, like exist in Python. They will generate compiler warnings, and will generate no assembly.
- Some code refactoring (type changes, for example)

+++ vm/SINVM.h
@@ -86,9 +86,9 @@ class SINVM
 	void execute_comparison(int reg_to_compare);
 	void execute_jmp();
 
-	// stack functions; ALWAYS use register A
-	void push_stack(int reg_to_push);
-	int pop_stack();
+	// stack functions
+	void push_stack(uint16_t reg_to_push);
+	uint16_t pop_stack();
 
 	// syscall utility
 	void free_heap_memory();
